GNU Make compiled to Wasm, and added to project.
Interesting idea, to add GNU Make (Used by micropythons build process) to the project. Running make on its own yields terrible results as it goes looking for binaries in /usr/bin, where none exist. But Make has a --dry-run flag which so far seems to be returning more or less the entire list of commands needed start to finish. This is also easy to get into JS, so we can then do .run(), parse the commands a little and hopefully bring us a few steps closer to a build.