Now using pythonsed package, to replicate sed commands, compilation of C files seems to be going fine, still need to handle mpy-cross
Pythonsed used in root.pack.br, to replicate UNIX sed. This was the final step of the micropython prerequisites before we can compile all the C files. This has now been done and the hex file fed back to the user, however for now it doesnt work, lots of things that could be doing this. In big need of a clean up so may do this before approaching mpy-cross