Mpy-cross compiled to wasm now added to project. We are now essentially building all of micropython from source, and linking with a CODAL cached buil.
Root.pack.br is getting quite large now so it may be time soon to try and minimise this. Possibly re-architect this whole system now we have access to python. UI also needs to actually allow the users to edit code, but for a demo this is fine for now.